+++ src/components/Editor/Runners/HtmlRunner/HtmlRunner.jsx
@@ -27,7 +27,6 @@ import RunnerControls from "../../../RunButton/RunnerControls";
 import { MOBILE_MEDIA_QUERY } from "../../../../utils/mediaQueryBreakpoints";
 
 function HtmlRunner() {
-  const webComponent = useSelector((state) => state.editor.webComponent);
   const project = useSelector((state) => state.editor.project);
   const projectCode = project.components;
   const projectImages = project.image_list;
@@ -62,7 +61,7 @@ function HtmlRunner() {
       (component) => `${component.name}.${component.extension}` === fileName,
     )[0];
 
-  const previewable = (file) => file.endsWith(".html");
+  const previewable = (file = "") => file.endsWith(".html");
   let defaultPreviewFile = "index.html";
 
   const pageExists = (page) => {
@@ -283,6 +282,8 @@ function HtmlRunner() {
           projectFile.content,
           mimeTypes.lookup(`${projectFile.name}.${projectFile.extension}`),
         );
+      } else if (matchingRegexes(allowedExternalLinks, srcNode.attrs[attr])) {
+        src = srcNode.attrs[attr];
       }
       srcNode.setAttribute(attr, src);
       srcNode.setAttribute("crossorigin", true);
@@ -325,7 +326,7 @@ function HtmlRunner() {
                   "output.preview",
                 )}`}</span>
               </Tab>
-              {!!!isEmbedded && !!!webComponent && (
+              {!!!isEmbedded && (
                 <a
                   className="btn btn--tertiary htmlrunner-link"
                   target="_blank"