+++ src/components/Editor/Runners/PyodideRunner/PyodideWorker.mock.js
@@ -1,3 +1,23 @@
-export const postMessage = (...args) => {
+export const postMessage = jest.fn((...args) => {
   console.log("Received postMessage with", args);
-};
+});
+
+class PyodideWorker {
+  constructor() {
+    this.postMessage = postMessage;
+    this.onmessage = null;
+    PyodideWorker.lastInstance = this;
+  }
+
+  postMessageFromWorker(message) {
+    if (this.onmessage) {
+      this.onmessage({ data: message });
+    }
+  }
+
+  static getLastInstance() {
+    return PyodideWorker.lastInstance;
+  }
+}
+
+export default PyodideWorker;