+++ src/containers/WebComponentLoader.jsx
@@ -1,6 +1,10 @@
 import React, { useEffect, useState } from "react";
 import { useSelector, useDispatch } from "react-redux";
-import { disableTheming, setSenseHatAlwaysEnabled } from "../redux/EditorSlice";
+import {
+  disableTheming,
+  setSenseHatAlwaysEnabled,
+  setLoadRemixDisabled,
+} from "../redux/EditorSlice";
 import WebComponentProject from "../components/WebComponentProject/WebComponentProject";
 import { useTranslation } from "react-i18next";
 import { setInstructions } from "../redux/InstructionsSlice";
@@ -17,33 +21,46 @@ import ToastCloseButton from "../utils/ToastCloseButton";
 
 import internalStyles from "../assets/stylesheets/InternalStyles.scss";
 import externalStyles from "../assets/stylesheets/ExternalStyles.scss";
+import editorStyles from "../assets/stylesheets/index.scss";
 import "../assets/stylesheets/Notifications.scss";
 import Style from "style-it";
+import { projectOwnerLoadedEvent } from "../events/WebComponentCustomEvents";
 
 const WebComponentLoader = (props) => {
   const {
+    assetsIdentifier,
     authKey,
     identifier,
     code,
     senseHatAlwaysEnabled = false,
     instructions,
     withProjectbar = false,
+    projectNameEditable = false,
     withSidebar = false,
     sidebarOptions = [],
     theme,
+    outputPanels = ["text", "visual"],
     embedded = false,
-    hostStyles,
+    hostStyles, // Pass in styles from the host
     showSavePrompt = false,
+    loadRemixDisabled = false,
+    outputOnly = false,
+    outputSplitView = false,
+    useEditorStyles = false, // If true use the standard editor styling for the web component
   } = props;
 
   const dispatch = useDispatch();
   const { t } = useTranslation();
   const [projectIdentifier, setProjectIdentifier] = useState(identifier);
   localStorage.setItem("authKey", authKey);
-  const user = useSelector((state) => state.auth.user);
+  const localStorageUser = authKey
+    ? JSON.parse(localStorage.getItem(authKey))
+    : null;
+  const user = useSelector((state) => state.auth.user || localStorageUser);
   const [loadCache, setLoadCache] = useState(!!!user);
   const [loadRemix, setLoadRemix] = useState(!!user);
   const project = useSelector((state) => state.editor.project);
+  const projectOwner = useSelector((state) => state.editor.project.user_name);
   const loading = useSelector((state) => state.editor.loading);
   const justLoaded = useSelector((state) => state.editor.justLoaded);
   const remixLoadFailed = useSelector((state) => state.editor.remixLoadFailed);
@@ -93,11 +110,18 @@ const WebComponentLoader = (props) => {
     }
   }, [loading, project]);
 
+  useEffect(() => {
+    if (justLoaded) {
+      document.dispatchEvent(projectOwnerLoadedEvent(projectOwner));
+    }
+  }, [projectOwner, justLoaded]);
+
   useProject({
     projectIdentifier: projectIdentifier,
+    assetsIdentifier: assetsIdentifier,
     code,
     accessToken: user?.access_token,
-    loadRemix,
+    loadRemix: loadRemix && !loadRemixDisabled,
     loadCache,
     remixLoadFailed,
   });
@@ -115,6 +139,10 @@ const WebComponentLoader = (props) => {
   }, [senseHatAlwaysEnabled, dispatch]);
 
   useEffect(() => {
+    dispatch(setLoadRemixDisabled(loadRemixDisabled));
+  }, [loadRemixDisabled, dispatch]);
+
+  useEffect(() => {
     if (instructions) {
       dispatch(setInstructions(instructions));
     }
@@ -129,7 +157,8 @@ const WebComponentLoader = (props) => {
         }}
       >
         <style>{externalStyles.toString()}</style>
-        <style>{hostStyles}</style>
+        {useEditorStyles && <style>{editorStyles.toString()}</style>}
+        {hostStyles && <style>{hostStyles}</style>}
         <Style>
           {internalStyles.toString()}
           <div id="wc" className={`--${cookies.theme || themeDefault}`}>
@@ -142,8 +171,12 @@ const WebComponentLoader = (props) => {
             />
             <WebComponentProject
               withProjectbar={withProjectbar}
+              nameEditable={projectNameEditable}
               withSidebar={withSidebar}
               sidebarOptions={sidebarOptions}
+              outputOnly={outputOnly}
+              outputPanels={outputPanels}
+              outputSplitView={outputSplitView}
             />
             {errorModalShowing && <ErrorModal />}
             {newFileModalShowing && <NewFileModal />}
@@ -154,7 +187,7 @@ const WebComponentLoader = (props) => {
     </>
   ) : (
     <>
-      <p>{t("webComponent.loading")}</p>;
+      <p>{t("webComponent.loading")}</p>
     </>
   );
 };