+++ src/events/WebComponentCustomEvents.js
@@ -8,6 +8,12 @@ const webComponentCustomEvent = (type, detail) =>
 
 export const codeChangedEvent = webComponentCustomEvent("editor-codeChanged");
 
+export const projectIdentifierChangedEvent = (detail) =>
+  webComponentCustomEvent("editor-projectIdentifierChanged", detail);
+
+export const projectOwnerLoadedEvent = (detail) =>
+  webComponentCustomEvent("editor-projectOwnerLoaded", detail);
+
 export const runCompletedEvent = (detail) =>
   webComponentCustomEvent("editor-runCompleted", detail);
 
@@ -21,3 +27,6 @@ export const logInEvent = webComponentCustomEvent("editor-logIn");
 export const signUpEvent = webComponentCustomEvent("editor-signUp");
 
 export const quizReadyEvent = webComponentCustomEvent("editor-quizReady");
+
+export const themeUpdatedEvent = (detail) =>
+  webComponentCustomEvent("editor-themeUpdated", detail);