+++ src/redux/EditorSlice.js
@@ -12,19 +12,24 @@ import {
   createRemix,
   deleteProject,
   readProjectList,
+  loadAssets,
 } from "../utils/apiCallHandler";
 
 export const syncProject = (actionName) =>
   createAsyncThunk(
     `editor/${actionName}Project`,
     async (
-      { project, identifier, locale, accessToken, autosave },
+      { project, identifier, locale, accessToken, autosave, assetsOnly },
       { rejectWithValue },
     ) => {
       let response;
       switch (actionName) {
         case "load":
-          response = await readProject(identifier, locale, accessToken);
+          if (assetsOnly) {
+            response = await loadAssets(identifier, locale, accessToken);
+          } else {
+            response = await readProject(identifier, locale, accessToken);
+          }
           break;
         case "loadRemix":
           response = await loadRemix(identifier, accessToken);
@@ -97,6 +102,7 @@ export const EditorSlice = createSlice({
     codeHasBeenRun: false,
     drawTriggered: false,
     isEmbedded: false,
+    isOutputOnly: false,
     browserPreview: false,
     isSplitView: true,
     isThemeable: true,
@@ -111,6 +117,7 @@ export const EditorSlice = createSlice({
     lastSavedTime: null,
     senseHatAlwaysEnabled: false,
     senseHatEnabled: false,
+    loadRemixDisabled: false,
     accessDeniedNoAuthModalShowing: false,
     accessDeniedWithAuthModalShowing: false,
     betaModalShowing: false,
@@ -124,6 +131,7 @@ export const EditorSlice = createSlice({
     deleteProjectModalShowing: false,
     sidebarShowing: true,
     modals: {},
+    errorDetails: {},
   },
   reducers: {
     closeFile: (state, action) => {
@@ -187,6 +195,9 @@ export const EditorSlice = createSlice({
     setEmbedded: (state, _action) => {
       state.isEmbedded = true;
     },
+    setIsOutputOnly: (state, action) => {
+      state.isOutputOnly = action.payload;
+    },
     setBrowserPreview: (state, _action) => {
       state.browserPreview = true;
     },
@@ -222,6 +233,9 @@ export const EditorSlice = createSlice({
     },
     setSenseHatEnabled: (state, action) => {
       state.senseHatEnabled = action.payload;
+    },
+    setLoadRemixDisabled: (state, action) => {
+      state.loadRemixDisabled = action.payload;
     },
     triggerDraw: (state) => {
       state.drawTriggered = true;
@@ -363,6 +377,9 @@ export const EditorSlice = createSlice({
     disableTheming: (state) => {
       state.isThemeable = false;
     },
+    setErrorDetails: (state, action) => {
+      state.errorDetails = action.payload;
+    },
   },
   extraReducers: (builder) => {
     builder.addCase("editor/saveProject/pending", (state) => {
@@ -392,8 +409,10 @@ export const EditorSlice = createSlice({
     });
     builder.addCase("editor/remixProject/pending", (state, action) => {
       state.saving = "pending";
+      state.saveTriggered = false;
     });
     builder.addCase("editor/remixProject/fulfilled", (state, action) => {
+      localStorage.removeItem(state.project.identifier);
       state.lastSaveAutosave = false;
       state.saving = "success";
       state.project = action.payload.project;
@@ -450,6 +469,7 @@ export const {
   setFocussedFileIndex,
   setPage,
   setEmbedded,
+  setIsOutputOnly,
   setBrowserPreview,
   setError,
   setIsSplitView,
@@ -459,6 +479,7 @@ export const {
   setProject,
   setSenseHatAlwaysEnabled,
   setSenseHatEnabled,
+  setLoadRemixDisabled,
   stopCodeRun,
   stopDraw,
   triggerCodeRun,
@@ -491,6 +512,7 @@ export const {
   showSidebar,
   hideSidebar,
   disableTheming,
+  setErrorDetails,
 } = EditorSlice.actions;
 
 export default EditorSlice.reducer;