+++ CHANGELOG.md
@@ -4,6 +4,120 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- RPi logo and text link at top left opens in new tab (#830)
+
+### Added
+
+- `editor-projectOwnerChanged` custom event (#1065)
+
+### Changed
+
+- "test" workflow to trigger on pull_request branches rather than push (#1069)
+
+## [0.25.5] - 2024-08-08
+
+### Fixed
+
+- CSS fix for the bottom of Sidebar (#1061)
+- Attach Skulpt files to the `document.body` to avoid issues with the shadow DOM (#1062)
+- Small fix to ensure webpack is using the correct webSocketURL for live reloading (#1051)
+- Only register `editor-wc` once (#1052)
+
+## [0.25.4] - 2024-06-20
+
+### Fixed
+
+- Revert: Quick fix for expired access token problem in web component (#1048)
+
+## [0.25.3] - 2024-06-19
+
+### Fixed
+
+- Quick fix for expired access token problem in web component (#1046)
+
+## [0.25.2] - 2024-06-13
+
+### Added
+
+- Web component now supports using standard editor-ui styles if useEditorStyles is true
+
+### Fixed
+
+- Fix bug where `skulpt` runs the first component rather than `main.py` (#1039)
+- Fixing bug with `HTML` projects loading after remix if they have multiple `HTML files` (#1040)
+
+## [0.25.1] - 2024-06-11
+
+### Fixed
+
+- Wrap `WebComponentLoader` in a `BrowserRouter` to fix routing issues (#1035)
+
+## [0.25.0] - 2024-06-06
+
+### Added
+
+- Add `embedded` attribute to web component (#1030)
+- Add `output_split_view` attribute to web component (#1030)
+
+### Changed
+
+- Changes to web component behaviour to support use in embedded view in editor-standalone (#1030)
+
+### Fixed
+
+## [0.24.0] - 2024-06-03
+
+### Added
+
+- Add `project_name_editable` attribute to web component (#1009)
+- Fires custom event when the theme changes (#1015)
+- Add `output_only` attribute to web component (#1019 & originally #782)
+- Add `assets_identifier` attribute to web component (#1019 & originally #901)
+- Enhance `code` attribute on web component to override project main component content (#1019 & originally #901)
+- Add `runCode`, `stopCode` & `rerunCode` methods to web component (#1019 & originally #899)
+- Send error details in "editor-runCompleted" event (#1019 & originally #915)
+- Return error details to web component (#1019 & originally #915)
+- Add `output_panels` attribute to web component (#1019 & originally #909)
+
+### Changed
+
+- Remove unused `/embedded/projects/:identifier` route (#1013)
+
+### Fixed
+
+- Remove unused `REACT_APP_LOGIN_ENABLED` env var (#1006)
+- Fix infinite remix loop when `BYPASS_AUTH` set in `editor-api` (#1007)
+- Fixes for docker-compose.yml (#1008)
+- Fix deprecation warnings in GitHub Actions (#1011)
+- Removed unused `isEmbedded` param from `useProject` call in `EmbeddedViewer` (#1016)
+- Improvements to Cypress specs in CI (#1017)
+- Fix warnings and verbose output when starting Webpack Dev Server (#1018)
+- Add e2e spec for project remix behaviour in web component (#1020)
+- Fix initial value of `user` in `WebComponentLoader` (#1021)
+- Make `authKey` in e2e web component spec more realistic (#1022)
+- Remove unused `ComponentStore` (#1023)
+
+## [0.23.0] - 2024-05-09
+
+### Added
+
+- Support to enable embedding iframes in HTML projects from in-house domains (#985)
+- Dispatch event when project identifier changes, e.g. after project is remixed (#2830)
+- Add `load_remix_disabled` attribute to web component (#992)
+
+### Changed
+
+- Invalidate cached project when project is remixed (#1003)
+
+### Fixed
+
+- Unit tests for `pyodide` runner (#976)
+- Remove broken `format` script (#991)
+
 ## [0.22.2] - 2024-03-18
 
 ### Added
@@ -11,6 +125,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Ability to use `page` query parameter in the embedded viewer when not a browser preview (#958)
 - ASSETS_URL env var to allow assets to be served from R2 bucket
 - `webpackDevServer` allowed headers for Astro Pi (#967)
+- Tests for the `PyodideWorker` (#981)
 
 ### Changed
 
@@ -750,7 +865,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Events in Web Component indicating whether Mission Zero criteria have been met (#113)
 
-[unreleased]: https://github.com/RaspberryPiFoundation/editor-ui/compare/v0.22.2...HEAD
+[unreleased]: https://github.com/RaspberryPiFoundation/editor-ui/compare/v0.25.5...HEAD
+[0.25.5]: https://github.com/RaspberryPiFoundation/editor-ui/releases/tag/v0.25.5
+[0.25.4]: https://github.com/RaspberryPiFoundation/editor-ui/releases/tag/v0.25.4
+[0.25.3]: https://github.com/RaspberryPiFoundation/editor-ui/releases/tag/v0.25.3
+[0.25.2]: https://github.com/RaspberryPiFoundation/editor-ui/releases/tag/v0.25.2
+[0.25.1]: https://github.com/RaspberryPiFoundation/editor-ui/releases/tag/v0.25.1
+[0.25.0]: https://github.com/RaspberryPiFoundation/editor-ui/releases/tag/v0.25.0
+[0.24.0]: https://github.com/RaspberryPiFoundation/editor-ui/releases/tag/v0.24.0
+[0.23.0]: https://github.com/RaspberryPiFoundation/editor-ui/releases/tag/v0.23.0
 [0.22.2]: https://github.com/RaspberryPiFoundation/editor-ui/releases/tag/v0.22.2
 [0.22.1]: https://github.com/RaspberryPiFoundation/editor-ui/releases/tag/v0.22.1
 [0.22.0]: https://github.com/RaspberryPiFoundation/editor-ui/releases/tag/v0.22.0